chore: moved home contents to contents/home/
